+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/application/App.kt
@@ -51,8 +51,8 @@ class App : Application() {
         ToggleService.init(
             applicationContext,
             mapOf(
-                ToggleService.TOGGLE_KEY_MESSAGE_COMPOSER to true,
-                ToggleService.TOGGLE_KEY_SOCKET_REFACTOR to true,
+                ToggleService.TOGGLE_KEY_MESSAGE_COMPOSER to false,
+                ToggleService.TOGGLE_KEY_SOCKET_REFACTOR to false,
             ),
         )
     }